Move deaggregator files into a subfolder called deaggregator for a bit more logical separation for imports
